requirements-mlops.txt
finrl>=0.3.7
pyqlib>=0.9.2
lightgbm>=4.0.0
stable-baselines3>=2.2.0
torch>=2.1.0
gymnasium>=0.29.0
scikit-learn>=1.3.0
